ARM-TEMPLATES/A10-vThunder_ADC-3NIC-VMSS/AZURE_FUNCTIONS/GetMetrics/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-common==1.1.28
azure-core==1.24.1
azure-functions==1.11.2
azure-identity==1.10.0
azure-mgmt-automation==1.0.0
azure-mgmt-compute==27.1.0
azure-mgmt-core==1.3.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
idna==3.3
isodate==0.6.1
msal==1.18.0
msal-extensions==1.0.0
msrest==0.7.1
oauthlib==3.2.0
portalocker==2.4.0
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.20.0
requests==2.28.0
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.2.0
urllib3==1.26.9